Add whatToDoNext method.
This function lives inside the getResultsInner recursive
function, and is created to avoid repeating what to
do next after a complete iteration or a failed iteration.
The logic is the same.

The end result is not readable of the recursive function
is not readable, though. Maybe a plain for loop will
be better.